--- a/src/main.rs
+++ b/src/main.rs
@@ -1,9 +1,11 @@
-pub mod solver;
-pub mod tools;
-use crate::solver::solve_puzzle;
-use crate::tools::puzzle_from_string;
 use std::fs;
 use std::time::Instant;
+
+use crate::solver::{solve_puzzle, Board, Move, State};
+use crate::tools::puzzle_from_string;
+
+pub mod solver;
+pub mod tools;
 
 fn main() {
     test1000();
@@ -12,34 +14,39 @@ fn main() {
 fn test1000() {
     const FILE_NAME: &str = "test_input/maps_moves.txt";
     let input = fs::read_to_string(FILE_NAME).expect("File not found.");
-    let mut i = 0;
-
     let before = Instant::now();
-    for line in input.lines() {
-        i += 1;
-        if i % 1 == 20 {
-            let now = Instant::now();
-            let (board, state) = puzzle_from_string(line);
-            let (_, _, _) = solve_puzzle(&board, state).unwrap();
-            let ms = now.elapsed().as_micros();
-            println!(
-                "Puzzle {}, solution found in: {}.{}ms",
-                i,
-                ms / 1000,
-                ms % 1000
-            );
-        } else {
-            let (board, state) = puzzle_from_string(line);
-            let (_, _, _) = solve_puzzle(&board, state).unwrap();
+    let (send, tasks) = crossbeam::channel::unbounded();
+    let mut received = 0;
+    rayon::scope(|handle| {
+        let mut submitted = 0;
+        for line in input.lines() {
+            let line = line.to_string();
+            submitted += 1;
+            let s_c = send.clone();
+            handle.spawn(move |_| {
+                let res = solve_board(line);
+                s_c.send(res).unwrap();
+            });
         }
-    }
+        while received < submitted {
+            let _next = tasks.recv().unwrap().unwrap();
+            received += 1;
+        }
+    });
+
     let ms = before.elapsed().as_micros();
     println!(
         "All {} solutions found in: {}.{}ms",
-        i,
+        received,
         ms / 1000,
         ms % 1000
     );
+}
+
+fn solve_board(line: String) -> Option<(Board, State, Vec<Move>)> {
+    let (board, state) = puzzle_from_string(&line);
+    let (_, state, moves) = solve_puzzle(&board, state)?;
+    Some((board, state, moves))
 }
 
 fn _test_diff_size() {