First commit uses heapless instead of vec, pulls in `heapless` as a dependency. 
Since it's doing a lot of allocations in a loop, and the max size is 12, we can put it on the stack instead of allocating each time, which `PieceType * Direction` `(3 * 4)`.

Just that gives a pretty good speed up from about 11.5 to 8.7 seconds. 

Second commit just runs all tests in parellell using `rayon`, I don't know if that's something you had in mind, or if you just wanted to improve single board solving speed, but that brings it down from 8.7 to 2 seconds for a run of the 995 test boards.